requirements.txt
annotated-types==0.7.0
anyio==4.9.0
argcomplete==3.6.2
asttokens==3.0.0
asyncio==3.4.3
bandit==1.8.3
beautifulsoup4==4.13.4
black==25.1.0
build==1.2.2.post1
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cleo==2.1.0
click==8.2.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
crashtest==0.4.1
cryptography==45.0.3
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
distlib==0.3.9
distro==1.9.0
docopt==0.6.2
dulwich==0.22.8
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.18.0
findpython==0.6.3
fonttools==4.58.0
fsspec==2024.10.0
# Editable Git install with no remote (gemini-audio-ai==0.1.0)
-e d:\onedrive\desktop\geminiendpoint\geminiaudioai\main
google-api-core==2.25.1
google-api-python-client==2.174.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-genai==1.18.0
googleapis-common-protos==1.70.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
installer==0.7.0
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
isort==6.0.1
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.10.0
joblib==1.5.1
Js2Py==0.74
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.8.1
keyring==25.6.0
kiwisolver==1.4.8
langchain==0.3.26
langchain-core==0.3.66
langchain-openai==0.3.27
langchain-text-splitters==0.3.8
langsmith==0.4.4
mando==0.7.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
memory-profiler==0.61.0
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.0
mss==10.0.0
mypy==1.15.0
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.6
oauthlib==3.3.1
openai==1.93.0
opencv-python==4.11.0.86
orjson==3.10.18
packaging==24.2
pandas==2.3.0
parso==0.8.4
pathspec==0.12.1
pbr==6.1.1
pbs-installer==2025.6.6
pexpect==4.9.0
pillow==11.2.1
pipreqs==0.4.13
pipx==1.7.1
pkginfo==1.12.1.2
platformdirs==4.3.8
pluggy==1.6.0
poetry==2.1.3
poetry-core==2.1.3
poetry-plugin-export==1.9.0
poetry-plugin-shell==1.0.1
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
pyjsparser==2.7.1
pyparsing==3.2.3
PyPrind==2.11.3
pyproject_hooks==1.2.0
pySmartDL==1.3.4
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pywin32==310
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyzmq==26.4.0
radon==6.0.1
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rsa==4.9.1
ruff==0.11.10
scikit-learn==1.7.0
scipy==1.15.3
setuptools==80.7.1
shellingham==1.5.4
six==1.17.0
snakeviz==2.2.2
sniffio==1.3.1
sounddevice==0.5.2
soupsieve==2.7
sourcery==1.36.0
SQLAlchemy==2.0.41
stack-data==0.6.3
stevedore==5.4.1
sympy==1.13.3
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tomlkit==0.13.3
tornado==6.5
tqdm==4.67.1
traitlets==5.14.3
trove-classifiers==2025.5.9.12
types-cachetools==6.0.0.20250525
types-cffi==1.17.0.20250523
types-defusedxml==0.7.0.20250516
types-docutils==0.21.0.20250526
types-pyasn1==0.6.0.20250516
types-pyaudio==0.2.16.20240516
types-Pygments==2.19.0.20250516
types-pyOpenSSL==24.1.0.20240722
types-requests==2.32.0.20250602
types-setuptools==80.9.0.20250529
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.4.0
userpath==1.9.2
virtualenv==20.31.2
wcwidth==0.2.13
websockets==15.0.1
yarg==0.1.10
zstandard==0.23.0
